+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add instances for more `time` types: `LocalTime`, `ZonedTime`, `UTCTime` and `NominalDiffTime`
+
 0.2.1
 ---